Fixes #1
This is a partial fix, as it solves it for 99% of the advisories.
There are a few advisories, where the CVE is referenced multiple
times across multiple advisories (in the same OS version) and thus
looking up the CVE in the online JSON files gives you multiple
fixed package versions. This only happens for a handful of cases
so we don't mind yet!